test: Text Tokens: Finding out sensible token limits
As BERT has a limitation of 512 input tokens and we
want to use it as cross encoder, the combined texts
of vehicle information and search query shouldn't
exceed 512 tokens. For that, we needed to find out
if we need to shorten our texts and how much to
shorten them. This notebook will test that.