test: generate queries with small LM inferences running on BHT cluster
Models as small as 7B parameters at the end didn't work all too well
for realistic query generation. So after this attempt, we decided to try
out the ChatGPT API in order to generate a huge amount of queries in
order to achieve a higher quality than we already have.