fix: generated queries (some manual data cleansing)
We accidentally used not the final vehicle information dataset but the
previous - 'truncated' one for the chatgpt query generation (which did
cost us 30$) - some information in it like vehicle no. and kba shouldn't
be used for the questions (as no one would state that they are interested
in exactly 'the vehicle with number XY'). So there was some manual cleanup
necessary.